+++ components/forms/controls/DateTime.jsx
@@ -1,61 +1,37 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-import DateTimePicker from 'react-datetime';
+import React from 'react';
+import MuiInput from '../base-controls/MuiInput';
 import { registerComponent } from 'meteor/vulcan:core';
+import withStyles from '@material-ui/core/styles/withStyles';
 
 
-class DateTime extends PureComponent {
+export const styles = theme => ({
   
-  constructor(props) {
-    super(props);
-    this.updateDate = this.updateDate.bind(this);
-  }
-
-  // when the datetime picker has mounted, SmartForm will catch the date value (no formsy mixin in this component)
-  componentDidMount() {
-    if (this.props.value) {
-      this.updateDate(this.props.value);
-    }
-  }
-
-  updateDate(date) {
-    this.context.updateCurrentValues({[this.props.name]: date});
-  }
-
-  render() {
-
-    const date = this.props.value ? (typeof this.props.value === 'string' ? new Date(this.props.value) : this.props.value) : null;
+  '@global': {
+    'input[type=datetime]::-ms-clear, input[type=datetime]::-ms-reveal': {
+      display: 'none',
+      width: 0,
+      height: 0,
+    },
+    'input[type=datetime]::-webkit-search-cancel-button': {
+      display: 'none',
+      '-webkit-appearance': 'none',
+    },
+    'input[type="datetime"]::-webkit-clear-button': {
+      display: 'none',
+      '-webkit-appearance': 'none',
+    },
     
-    return (
-      <div className="form-group row">
-        <label className="control-label col-sm-3">{this.props.label}</label>
-        <div className="col-sm-9">
-          <DateTimePicker
-            value={date}
-            // newDate argument is a Moment object given by react-datetime
-            onChange={newDate => this.updateDate(newDate._d)}
-            format={"x"}
-            inputProps={{name: this.props.name}}
-          />
-        </div>
-      </div>
-    );
-  }
-}
+    'input[type="datetime"]::-webkit-inner-spin-button,input[type="datetime"]::-webkit-outer-spin-button': {
+      '-webkit-appearance': 'none',
+      margin: 0,
+    },
+  },
+  
+});
 
-DateTime.propTypes = {
-  control: PropTypes.any,
-  datatype: PropTypes.any,
-  group: PropTypes.any,
-  label: PropTypes.string,
-  name: PropTypes.string,
-  value: PropTypes.any,
-};
 
-DateTime.contextTypes = {
-  updateCurrentValues: PropTypes.func,
-};
+const DateTimeComponent = ({ refFunction, classes, ...properties }) =>
+  <MuiInput {...properties} ref={refFunction} type="datetime-local"/>;
 
-export default DateTime;
 
-registerComponent('FormComponentDateTime', DateTime);
+registerComponent('FormComponentDateTime', DateTimeComponent, [withStyles, styles]);