+++ components/forms/controls/Time.jsx
@@ -1,73 +1,37 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-import DateTimePicker from 'react-datetime';
+import React from 'react';
+import MuiInput from '../base-controls/MuiInput';
 import { registerComponent } from 'meteor/vulcan:core';
+import withStyles from '@material-ui/core/styles/withStyles';
 
-class Time extends PureComponent {
+
+export const styles = theme => ({
   
-  constructor(props) {
-    super(props);
-    this.updateDate = this.updateDate.bind(this);
-  }
+  '@global': {
+    'input[type=time]::-ms-clear, input[type=time]::-ms-reveal': {
+      display: 'none',
+      width: 0,
+      height: 0,
+    },
+    'input[type=time]::-webkit-search-cancel-button': {
+      display: 'none',
+      '-webkit-appearance': 'none',
+    },
+    'input[type="time"]::-webkit-clear-button': {
+      display: 'none',
+      '-webkit-appearance': 'none',
+    },
+    
+    'input[type="time"]::-webkit-inner-spin-button,input[type="time"]::-webkit-outer-spin-button': {
+      '-webkit-appearance': 'none',
+      margin: 0,
+    },
+  },
+  
+});
 
-  // when the datetime picker has mounted, SmartForm will catch the date value (no formsy mixin in this component)
-  // componentDidMount() {
-  //   if (this.props.value) {
-  //     this.context.updateCurrentValues({[this.props.path]: this.props.value});
-  //   }
-  // }
 
-  updateDate(mDate) {
-    // if this is a properly formatted moment date, update time
-    if (typeof mDate === 'object') {
-      this.context.updateCurrentValues({[this.props.path]: mDate.format('HH:mm')});
-    }
-  }
+const TimeComponent = ({ refFunction, classes, ...properties }) =>
+  <MuiInput {...properties} ref={refFunction} type="time"/>;
 
-  render() {
 
-    const date = new Date();
-
-    // transform time string into date object to work inside datetimepicker
-    const time = this.props.value;
-    if (time) {
-      date.setHours(parseInt(time.substr(0,2)), parseInt(time.substr(3,5)));
-    } else {
-      date.setHours(0,0);
-    }
-
-    return (
-      <div className="form-group row">
-        <label className="control-label col-sm-3">{this.props.label}</label>
-        <div className="col-sm-9">
-          <DateTimePicker
-            value={date}
-            viewMode="time"
-            dateFormat={false}
-            timeFormat="HH:mm"
-            // newDate argument is a Moment object given by react-datetime
-            onChange={newDate => this.updateDate(newDate)}
-            inputProps={{name: this.props.name}}
-          />
-        </div>
-      </div>
-    );
-  }
-}
-
-Time.propTypes = {
-  control: PropTypes.any,
-  datatype: PropTypes.any,
-  group: PropTypes.any,
-  label: PropTypes.string,
-  name: PropTypes.string,
-  value: PropTypes.any,
-};
-
-Time.contextTypes = {
-  updateCurrentValues: PropTypes.func,
-};
-
-export default Time;
-
-registerComponent('FormComponentTime', Time);
+registerComponent('FormComponentTime', TimeComponent, [withStyles, styles]);