+++ components/index.js
@@ -30,7 +30,9 @@ import './forms/controls/Checkbox';
 import './forms/controls/CheckboxGroup';
 import './forms/controls/CountrySelect';
 import './forms/controls/Date';
+import './forms/controls/DateRdt';
 import './forms/controls/DateTime';
+import './forms/controls/DateTimeRdt';
 import './forms/controls/Default';
 import './forms/controls/Email';
 import './forms/controls/Number';
@@ -40,6 +42,7 @@ import './forms/controls/RegionSelect';
 import './forms/controls/Select';
 import './forms/controls/Textarea';
 import './forms/controls/Time';
+import './forms/controls/TimeRdt';
 import './forms/controls/Url';
 
 import './theme/ThemeStyles';