+++ modules/sampleTheme.js
@@ -11,7 +11,7 @@ import red from '@material-ui/core/colors/red';
  * Sample theme to get you out of the gate quickly
  *
  * For a complete list of configuration variables see:
- * https://material-ui-next.com/customization/themes/
+ * https://material-ui.com/customization/themes/
  *
  */